trible-backend/requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
numpy
requests
beautifulsoup4
python-dotenv
websockets
